Link to latest, not listed version
This patch tweaks read-related so that rather than linking to the
version where the symbol(s) are listed as related are listed, things
sharing the version of the thing for which related symbols are requested
will be returned.

Note that this may generate invlalid links, however it is likely to give
more current results and so is provided.