Reimplement thing->prior-versions via search
Search already has support for filtering on functions, so why duplicate
that logic? Just reimplement the bugged thing->prior-versions in terms
of the searching operation. In doing so it also fixes some weird bugs
with regards to the handling of 1.7.0-snapshot versions.

Fixes clojure-grimoire/grimoire#209

This commit also introduces some optimization to search. String and set
patterns will _not_ generate listing and equality scanning
operations. Previously the pattern `[:group "org.clojure"]` would
enumerate all groups, and scan for the groups such that the name of the
group was string equal to `"org.clojure"`. With this tweak in no
enumeration will occur, instead the Group thing naming `"org.clojure"`
will be constructed directly and its existance checked for.

Also adds the appropriate entries to the README and updates the README
for guten-tag output.