Add default reader function
In these modern times, strange Clojure objects aren't printed as

Add a function delegating to clojure.edn/read-string primed with a
default reader function to grimoire.util. Substitute the calls to
clojure.edn/read-string with calls to that function so that unknown tags
don't cause exceptions, but result in a marker data structure.

I'm not sure if this is a good implementation. – The choice of the
marker data structure is quite arbitrary and I'm not content with where
I put the new function, but I don't know a more sensible place either.