Start working on a "2.0" Grimoire Thing structure
This work is largely in response to martinklepsch/cljdoc#5 - I wanted
to try and get my thoughts around how I see future Grimoire-style
doctrees being structured and wanted to see if I could do so more
explicitly by abandoning guten-tag in favor of clojure.spec(.alpha).

The particular goal here is to start working towards a modularized
"thing" structure. The Group/Artifact/Version root used to address a
thing is more-or-less sacrosanct. In light of tools.deps and wanting
to use alternate resolution strategies some reconsideration for
addressing artifacts as (repository URL + SHA) etc. may be in order,
but that's out of scope for now.

On top of this basic root, I wanted to try and sit both clj* family
documentation trees which have a common (namespace, name) addressing
scheme, as well as Java documentation which has a (package, class,
name) addressing scheme. This is still a work in progress, but is
sketched at.

Documentation (or documents, notes was an awful name) and examples are
now better codified as being annotations to any versioned and
addressable entity, other annotations excepted.

My vision here is that Grimoire "2.0" stores will have a concrete
structure (an index) which enumerates the types/schemas of content in
the store and contains metadata records or URIs for every record in
the store.

A schema / structure for such stores needs to be established, and
perhaps some distinction should be made between a "package unique
identifier", a platform-specific entity identifier which lives in the
context of a package, and the identifier for some commentary about an
artifact in a packaged object.